+++ test/prediction.ts
@@ -16,10 +16,7 @@ class PredictionTest {
     public before () {
         should();
         use(chaiAsPromised);
-        this.fxn = new Function({
-            accessKey: process.env.FXN_ACCESS_KEY,
-            url: process.env.FXN_API_URL
-        });
+        this.fxn = new Function();
     }
 
     @mocha.test